implemented a* bathing for miners. Still need to implement for blobs, but can use same pather class.
NOTE: For some reason, the check to see if a node is in the closed set does not work. I need to fix this as it often results in an infinite loops instead of a miner finding its target.
Also, DFSPather is incomplete, does not work, is not in use, and may or may not ever be completed.